+++ play/index.php
@@ -169,7 +169,7 @@ if ($demo_id) {
  */
 if ($random) {
     if ($start < 1) {
-        $oid = Random::get_single_song($_REQUEST['type']);
+        $oid = Random::get_single_song($_REQUEST['random_type']);
         // Save this one in case we do a seek
         $_SESSION['random']['last'] = $oid;
     }